ViewModel.csが長くなったのでWebScraping.csに分離。
