+++ Scraping/Scraping/ViewModel.cs
@@ -5,9 +5,6 @@ using System.Text;
 using System.Threading.Tasks;
 using System.ComponentModel;
 using System.Collections.ObjectModel;
-using AngleSharp.Html.Dom;
-using AngleSharp.Html.Parser;
-using System.Net.Http;
 using System.Runtime.CompilerServices;
 using System.Windows.Input;
 
@@ -50,8 +47,8 @@ namespace Scraping
         /// <summary>
         /// データバインド用のリスト
         /// </summary>
-        private ObservableCollection<Book> _viewList;
-        public ObservableCollection<Book> ViewList
+        private List<Book> _viewList;
+        public List<Book> ViewList
         {
             get
             {
@@ -112,13 +109,13 @@ namespace Scraping
             {
                 HeaderText = "お気に入り表示";
                 ButtonText = "全表示に切り替え";
-                ViewList = new ObservableCollection<Book>(_favoriteList);
+                ViewList = new List<Book>(_favoriteList);
             }
             else
             {
                 HeaderText = "全表示";
                 ButtonText = "お気に入り表示に切り替え";
-                ViewList = new ObservableCollection<Book>(_bookList);
+                ViewList = new List<Book>(_bookList);
             }
         }
     }
@@ -155,87 +152,5 @@ namespace Scraping
         public string Title { get; set; }
         public string ReleaseDate { get; set; }
         public string Author { get; set; }
-    }
-    public class WebScraping
-    {
-        static HttpClient client = new HttpClient();
-
-        public static List<Book> GetWebData()
-        {
-            var urlstring = @"http://yurinavi.com/yuri-calendar/";
-            var document = default(IHtmlDocument);
-            using (var stream = client.GetStreamAsync(urlstring).Result)
-            {
-                var parser = new HtmlParser();
-                document = parser.ParseDocumentAsync(stream).Result;
-            }
-            var dates = document.QuerySelectorAll(@"td.column-1:not(#tablepress-152 > tbody > tr > td)");
-            var books = document.QuerySelectorAll(@"td.column-3");
-            var count = 0;
-            var data = new List<Book>();
-            foreach (var date in dates)
-            {
-                if (date.TextContent != "")
-                {
-                    if (date.TextContent.Substring(0, 1) != "▼")
-                    {
-                        var dateArray = date.TextContent.Split('\n');
-                        var releaseDate = dateArray[0] + "(" + dateArray[1] + ")";
-                        var bookData = books[count].TextContent.Split('\n');
-                        if (bookData.Length == 2)
-                        {
-                            var title = bookData[0];
-                            var author = bookData[1];
-                            data.Add(new Book()
-                            {
-                                ReleaseDate = releaseDate,
-                                Title = title,
-                                Author = author
-                            });
-                            count++;
-                        }
-                        else
-                        {
-                            var title = bookData[0];
-                            data.Add(new Book()
-                            {
-                                ReleaseDate = releaseDate,
-                                Title = title,
-                                Author = ""
-                            });
-                            count++;
-                        }
-                    }
-                }
-                else
-                {
-                    var bookData = books[count].TextContent.Split('\n');
-                    if (bookData.Length == 2)
-                    {
-                        var title = bookData[0];
-                        var author = bookData[1];
-                        data.Add(new Book()
-                        {
-                            ReleaseDate = "",
-                            Title = title,
-                            Author = author
-                        });
-                        count++;
-                    }
-                    else
-                    {
-                        var title = bookData[0];
-                        data.Add(new Book()
-                        {
-                            ReleaseDate = "",
-                            Title = title,
-                            Author = ""
-                        });
-                        count++;
-                    }
-                }
-            }
-            return data;
-        }
     }
 }
